rename : model, domain, entity에 해당하는 클래스 파일 폴더 위치 변경
