+++ src/main/java/racingcar/model/domain/CarsIterator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.domain;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
