+++ src/main/java/racingcar/model/io/RacingChance.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.io;
 
 import racingcar.utils.Constants;
